Device_Management/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
djangorestframework==3.14.0
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
